move func to cron handlers file
